Refactor package structure and naming conventions for consistency.
