+++ WeatherService/src/main/java/New_Foreflight/Weather/controller/WeatherServiceController.java
@@ -1,6 +1,6 @@
-package New_Foreflight.Weather.Controllers;
+package New_Foreflight.Weather.controller;
 
-import New_Foreflight.Weather.DTO.AirportWeatherResponse;
+import New_Foreflight.Weather.dto.AirportWeatherResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -9,14 +9,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import New_Foreflight.Weather.Service.Weatherservice;
+import New_Foreflight.Weather.service.WeatherService;
 
 @RestController
 @RequestMapping("/api/v1")
 public class WeatherServiceController {
 
     @Autowired
-    Weatherservice weatherService;
+    WeatherService weatherService;
 
     @GetMapping(value = "/getAirportWeather")
     public ResponseEntity<AirportWeatherResponse> getAirportWeather(@RequestParam String airportCode) {
